requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
httpx==0.26.0
python-dotenv==1.0.0
python-dateutil==2.8.2
bcrypt==4.0.1

